+++ app/src/main/java/com/mickstarify/zooforzotero/SyncSetup/ui/screens/SyncSetupLandingScreen.kt
@@ -1,7 +1,5 @@
 package com.mickstarify.zooforzotero.SyncSetup.ui.screens
 
-import android.content.Intent
-import android.net.Uri
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
@@ -35,6 +33,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.mickstarify.zooforzotero.R
 import com.mickstarify.zooforzotero.SyncSetup.ui.components.AuthenticationOptionsSection
+import com.mickstarify.zooforzotero.SyncSetup.ui.components.LegalDisclaimerSection
 import com.mickstarify.zooforzotero.SyncSetup.ui.components.WelcomeHeroSection
 import com.mickstarify.zooforzotero.SyncSetup.viewmodels.SyncSetupViewModel
 import com.mickstarify.zooforzotero.common.Constants
@@ -56,10 +55,8 @@ fun SyncSetupLandingScreen(
     ) {
         Spacer(modifier = Modifier.height(32.dp))
 
-        // Welcome Hero Section
         WelcomeHeroSection()
 
-        // Authentication Options
         AuthenticationOptionsSection(
             selectedOption = state.selectedSyncOption,
             onOptionSelected = { option ->
@@ -69,66 +66,7 @@ fun SyncSetupLandingScreen(
 
         Spacer(modifier = Modifier.height(32.dp))
 
-        // Disclaimer
-        Text(
-            text = stringResource(R.string.app_frontpage_disclaimer),
-            style = MaterialTheme.typography.bodyMedium,
-            color = MaterialTheme.colorScheme.onSurfaceVariant,
-            textAlign = TextAlign.Center,
-            modifier = Modifier.padding(horizontal = 16.dp)
-        )
-
-        Spacer(modifier = Modifier.height(8.dp))
-
-        // Legal Agreement Text
-        val annotatedText = buildAnnotatedString {
-            append(stringResource(R.string.legal_agreement_prefix))
-
-            pushStringAnnotation(tag = "terms", annotation = Constants.Legal.TERMS_OF_USE_URL)
-            withStyle(
-                style = SpanStyle(
-                    color = MaterialTheme.colorScheme.primary,
-                    textDecoration = TextDecoration.Underline
-                )
-            ) {
-                append(stringResource(R.string.terms_of_use))
-            }
-            pop()
-
-            append(stringResource(R.string.legal_agreement_middle))
-
-            pushStringAnnotation(tag = "privacy", annotation = Constants.Legal.PRIVACY_POLICY_URL)
-            withStyle(
-                style = SpanStyle(
-                    color = MaterialTheme.colorScheme.primary,
-                    textDecoration = TextDecoration.Underline
-                )
-            ) {
-                append(stringResource(R.string.privacy_policy))
-            }
-            pop()
-        }
-
-        ClickableText(
-            text = annotatedText,
-            style = MaterialTheme.typography.bodyMedium.copy(
-                color = MaterialTheme.colorScheme.onSurfaceVariant,
-                textAlign = TextAlign.Center
-            ),
-            modifier = Modifier.padding(horizontal = 16.dp),
-            onClick = { offset ->
-                annotatedText.getStringAnnotations(tag = "terms", start = offset, end = offset)
-                    .firstOrNull()?.let { annotation ->
-                        val intent = Intent(Intent.ACTION_VIEW, Uri.parse(annotation.item))
-                        context.startActivity(intent)
-                    }
-                annotatedText.getStringAnnotations(tag = "privacy", start = offset, end = offset)
-                    .firstOrNull()?.let { annotation ->
-                        val intent = Intent(Intent.ACTION_VIEW, Uri.parse(annotation.item))
-                        context.startActivity(intent)
-                    }
-            }
-        )
+        LegalDisclaimerSection()
 
         Spacer(modifier = Modifier.height(32.dp))
 
@@ -148,52 +86,6 @@ fun SyncSetupLandingScreen(
 
         Spacer(modifier = Modifier.height(24.dp))
     }
-
-    // API Key Dialog (Legacy - should be removed once navigation is fully implemented)
-    if (state.showApiKeyDialog) {
-        ApiKeyDialog(
-            onSubmit = { apiKey ->
-                onEvent(SyncSetupViewModel.Event.SubmitApiKey(apiKey))
-            },
-            onDismiss = {
-                onEvent(SyncSetupViewModel.Event.DismissApiKeyDialog)
-            }
-        )
-    }
-}
-
-@Composable
-private fun ApiKeyDialog(
-    onSubmit: (String) -> Unit,
-    onDismiss: () -> Unit
-) {
-    var apiKey by remember { mutableStateOf("") }
-
-    AlertDialog(
-        onDismissRequest = onDismiss,
-        title = { Text("Enter your API Key") },
-        text = {
-            OutlinedTextField(
-                value = apiKey,
-                onValueChange = { apiKey = it },
-                label = { Text("API Key") },
-                singleLine = true
-            )
-        },
-        confirmButton = {
-            TextButton(
-                onClick = { onSubmit(apiKey) },
-                enabled = apiKey.isNotBlank()
-            ) {
-                Text("Submit")
-            }
-        },
-        dismissButton = {
-            TextButton(onClick = onDismiss) {
-                Text("Cancel")
-            }
-        }
-    )
 }
 
 @Preview
@@ -212,7 +104,7 @@ private fun SyncSetupLandingScreenPreview() {
 private fun SyncSetupLandingScreenWithApiDialogPreview() {
     ZoteroTheme {
         SyncSetupLandingScreen(
-            state = SyncSetupViewModel.State(showApiKeyDialog = true),
+            state = SyncSetupViewModel.State(),
             onEvent = {}
         )
     }
